Ignore player position updates when the player is not present in the player list
New ws-server code sends a batch of _all_ position updates every X time, so the
mod will be receiving position updates for the player's self as well as all the
other players. Since we don't expect the player to be in the list table, only
the ghosts, guard against a failure here.