Fix PlayerPositions flushing
Serialized proto data concatenation behaves the same as merging the
deserialized data. By being deliberate about our message structure,
we can create data that can be concatenated and decoded on the other
side as though it was a single message.

Added a couple tests to verify the behavior is correct.