Replace 'skip_heart' method with 'fake_orb' tagged orb-hearts
Instead of incrementing/decrementing a 'skip_heart' variable,
add 'fake_orb' tags to hearts spawned by orbs. If this tag exists
on a picked-up heart, it will not be transmitted to other players
and not count for stats, as before. This fixes unpicked and cursed
orb hearts canceling out later heart pickups